Extract onRowEditStop helpers and consolidate DataGrid helper files
