+++ moped-editor/src/components/DataGridPro/DataGridActions.js
@@ -11,7 +11,7 @@ import {
   DeleteOutline as DeleteOutlineIcon,
 } from "@mui/icons-material";
 
-import { defaultEditColumnIconStyle } from "src/utils/dataGridHelpers";
+import { defaultEditColumnIconStyle } from "src/components/DataGridPro/utils/helpers.js";
 
 /** Component for Data Grid table action buttons
  * @param {Number} id - Data Grid row id (same as project id)
