--- moped-editor/src/utils/dataGridHelpers.js
@@ -1,16 +1,0 @@
-import { GridRowEditStopReasons } from "@mui/x-data-grid-pro";
-
-export const defaultEditColumnIconStyle = { fontSize: "24px" };
-
-export const handleRowEditStop = (rows, setRows) => (params, event) => {
-  const id = params.row.id;
-  if (params.reason === GridRowEditStopReasons.enterKeyDown) {
-    event.defaultMuiPrevented = true;
-  }
-  // Delete new row if user escapes out of edit mode
-  if (params.reason === GridRowEditStopReasons.escapeKeyDown) {
-    if (params.row.isNew) {
-      setRows(rows.filter((row) => row.id !== id));
-    }
-  }
-};
