+++ moped-editor/src/views/dashboard/UserSavedViewsTable.js
@@ -17,7 +17,7 @@ import DataGridActions from "src/components/DataGridPro/DataGridActions";
 import DataGridTextField from "src/components/DataGridPro/DataGridTextField";
 import DeleteConfirmationModal from "../projects/projectView/DeleteConfirmationModal";
 import { FormattedDateString } from "src/utils/dateAndTime";
-import { handleRowEditStop } from "src/utils/dataGridHelpers";
+import { handleRowEditStop } from "src/components/DataGridPro/utils/helpers.js";
 
 const requiredFields = ["description"];
