+++ moped-editor/src/views/projects/projectView/ProjectFiles.js
@@ -30,7 +30,7 @@ import DataGridTextField from "src/components/DataGridPro/DataGridTextField";
 import ProjectFilesTypeSelect from "src/views/projects/projectView/ProjectFilesTypeSelect";
 import DeleteConfirmationModal from "src/views/projects/projectView/DeleteConfirmationModal";
 import DataGridActions from "src/components/DataGridPro/DataGridActions";
-import { handleRowEditStop } from "src/utils/dataGridHelpers";
+import { handleRowEditStop } from "src/components/DataGridPro/utils/helpers.js";
 import { useUser } from "src/auth/user";
 
 // reshape the array of file types into an object with key id, value name
